# auth-service/src/main/resources/openapi/auth-api.yaml
openapi: 3.1.2
info:
    title: HvostID Auth Service
    description: Authentication API - registration, login with opaque tokens
    version: 1.0.0

servers:
    -   url: http://localhost:8080
        description: API Gateway (local)
    -   url: http://localhost:8081
        description: Auth Service (direct)

paths:
    /api/v1/auth/register:
        post:
            tags:
                - Auth
            summary: Register a new user
            description: Creates a new user account with buyer role by default.
            operationId: register
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/RegisterRequest'
            responses:
                '201':
                    description: User created successfully
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/UserResponse'
                '400':
                    description: Validation error
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
                '409':
                    description: Email already registered
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'

    /api/v1/auth/login:
        post:
            tags:
                - Auth
            summary: Log in with email and password
            description: |
                Authenticates the user and returns opaque access and refresh tokens.
                Tokens are random strings stored in the database (not JWT).
            operationId: login
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/LoginRequest'
            responses:
                '200':
                    description: Login successful
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/LoginResponse'
                '400':
                    description: Validation error
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
                '401':
                    description: Invalid credentials
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'

components:
    schemas:
        RegisterRequest:
            type: object
            required:
                - email
                - password
                - name
            properties:
                email:
                    type: string
                    format: email
                    example: user@example.com
                password:
                    type: string
                    format: password
                    minLength: 8
                    maxLength: 128
                    example: mySecurePassword1
                name:
                    type: string
                    maxLength: 255
                    example: John Doe

        LoginRequest:
            type: object
            required:
                - email
                - password
            properties:
                email:
                    type: string
                    format: email
                    example: user@example.com
                password:
                    type: string
                    format: password
                    example: mySecurePassword1

        UserResponse:
            type: object
            properties:
                id:
                    type: integer
                    format: int64
                    example: 1
                email:
                    type: string
                    example: user@example.com
                name:
                    type: string
                    example: John Doe
                role:
                    type: string
                    enum: [ buyer, seller, moderator, admin ]
                    example: buyer

        LoginResponse:
            type: object
            properties:
                accessToken:
                    type: string
                    description: Opaque access token (Base64-encoded random bytes)
                    example: dGhpcyBpcyBhIHNhbXBsZSB0b2tlbg
                refreshToken:
                    type: string
                    description: Opaque refresh token (Base64-encoded random bytes)
                    example: cmVmcmVzaCB0b2tlbiBleGFtcGxl
                expiresIn:
                    type: integer
                    format: int64
                    description: Access token lifetime in seconds
                    example: 1800

        ErrorResponse:
            type: object
            properties:
                status:
                    type: integer
                    example: 400
                error:
                    type: string
                    example: Bad Request
                message:
                    type: string
                    example: email - Invalid email format
                timestamp:
                    type: string
                    format: date-time